# src/main/java/za/co/bts/words/sensitive/model/SensitiveWord.java
package za.co.bts.words.sensitive.model;

import jakarta.persistence.*;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.util.UUID;

@Data
@AllArgsConstructor
@NoArgsConstructor
@Entity
@Table(name = "sensitive_words", schema = "dbo")
public class SensitiveWord {
    @Id
    @GeneratedValue(strategy = GenerationType.UUID)
    private UUID id;

    @Column(nullable = false, unique = true)
    private String word;
}